Halfway thru cleanup and reorg
