+++ predicators/behavior_utils/__init__.py
@@ -1,0 +1,2 @@
+"""All BEHAVIOR-related code that doesn't directly pertain to the BEHAVIOR
+environment interface."""
