+++ predicators/envs/behavior_options.py
@@ -8,12 +8,12 @@ import numpy as np
 import scipy
 from numpy.random._generator import Generator
 
+from predicators.behavior_utils.behavior_utils import get_aabb_volume, \
+    get_closest_point_on_aabb
 from predicators.structs import Array, State
-from predicators.utils import get_aabb_volume, get_closest_point_on_aabb
 
 try:
     import pybullet as p
-    from igibson import object_states
     from igibson.envs.behavior_env import \
         BehaviorEnv  # pylint: disable=unused-import
     from igibson.external.pybullet_tools.utils import CIRCULAR_LIMITS, \
@@ -21,128 +21,12 @@ try:
     from igibson.object_states.on_floor import \
         RoomFloor  # pylint: disable=unused-import
     from igibson.objects.articulated_object import URDFObject
-    from igibson.robots.behavior_robot import \
-        BRBody  # pylint: disable=unused-import
-    from igibson.robots.robot_base import \
-        BaseRobot  # pylint: disable=unused-import
     from igibson.utils import sampling_utils
     from igibson.utils.behavior_robot_planning_utils import \
         plan_base_motion_br, plan_hand_motion_br
 
 except (ImportError, ModuleNotFoundError) as e:
     pass
-
-_ON_TOP_RAY_CASTING_SAMPLING_PARAMS = {
-    "max_angle_with_z_axis": 0.17,
-    "bimodal_stdev_fraction": 1e-6,
-    "bimodal_mean_fraction": 1.0,
-    "max_sampling_attempts": 50,
-    "aabb_offset": 0.01,
-}
-
-
-def get_body_ids(
-    env: "BehaviorEnv",
-    include_self: bool = False,
-    include_right_hand: bool = False,
-) -> List[int]:
-    """Function to return a list of body_ids for all objects in the scene for
-    collision checking depending on whether navigation or grasping/ placing is
-    being done."""
-    ids = []
-    for obj in env.scene.get_objects():
-        if isinstance(obj, URDFObject):
-            # We want to exclude the floor since we're always floating and
-            # will never practically collide with it, but if we include it
-            # in collision checking, we always seem to collide.
-            if obj.name != "floors":
-                ids.extend(obj.body_ids)
-
-    if include_self:
-        ids.append(env.robots[0].parts["left_hand"].get_body_id())
-        ids.append(env.robots[0].parts["body"].get_body_id())
-        ids.append(env.robots[0].parts["eye"].get_body_id())
-        if not include_right_hand:
-            ids.append(env.robots[0].parts["right_hand"].get_body_id())
-
-    return ids
-
-
-def detect_collision(bodyA: int, object_in_hand: Optional[int] = None) -> bool:
-    """Detects collisions between bodyA in the scene (except for the object in
-    the robot's hand)"""
-    collision = False
-    for body_id in range(p.getNumBodies()):
-        if body_id in [bodyA, object_in_hand]:
-            continue
-        closest_points = p.getClosestPoints(bodyA, body_id, distance=0.01)
-        if len(closest_points) > 0:
-            collision = True
-            break
-    return collision
-
-
-def detect_robot_collision(robot: "BaseRobot") -> bool:
-    """Function to detect whether the robot is currently colliding with any
-    object in the scene."""
-    object_in_hand = robot.parts["right_hand"].object_in_hand
-    return (detect_collision(robot.parts["body"].body_id)
-            or detect_collision(robot.parts["left_hand"].body_id)
-            or detect_collision(robot.parts["right_hand"].body_id,
-                                object_in_hand))
-
-
-def reset_and_release_hand(env: "BehaviorEnv") -> None:
-    """Resets the state of the right hand."""
-    env.robots[0].set_position_orientation(env.robots[0].get_position(),
-                                           env.robots[0].get_orientation())
-    for _ in range(50):
-        env.robots[0].parts["right_hand"].set_close_fraction(0)
-        env.robots[0].parts["right_hand"].trigger_fraction = 0
-        p.stepSimulation()
-
-
-def get_delta_low_level_base_action(robot_z: float,
-                                    original_orientation: Tuple,
-                                    old_xytheta: Array, new_xytheta: Array,
-                                    action_space_shape: Tuple) -> Array:
-    """Given a base movement plan that is a series of waypoints in world-frame
-    position space, convert pairs of these points to a base movement action in
-    velocity space.
-
-    Note that we cannot simply subtract subsequent positions because the
-    velocity action space used by BEHAVIOR is not defined in the world
-    frame, but rather in the frame of the previous position.
-    """
-    ret_action = np.zeros(action_space_shape, dtype=np.float32)
-
-    # First, get the old and new position and orientation in the world
-    # frame as numpy arrays
-    old_pos = np.array([old_xytheta[0], old_xytheta[1], robot_z])
-    old_orn_quat = p.getQuaternionFromEuler(
-        np.array(
-            [original_orientation[0], original_orientation[1],
-             old_xytheta[2]]))
-    new_pos = np.array([new_xytheta[0], new_xytheta[1], robot_z])
-    new_orn_quat = p.getQuaternionFromEuler(
-        np.array(
-            [original_orientation[0], original_orientation[1],
-             new_xytheta[2]]))
-
-    # Then, simply get the delta position and orientation by multiplying the
-    # inverse of the old pose by the new pose
-    inverted_old_pos, inverted_old_orn_quat = p.invertTransform(
-        old_pos, old_orn_quat)
-    delta_pos, delta_orn_quat = p.multiplyTransforms(inverted_old_pos,
-                                                     inverted_old_orn_quat,
-                                                     new_pos, new_orn_quat)
-
-    # Finally, convert the orientation back to euler angles from a quaternion
-    delta_orn = p.getEulerFromQuaternion(delta_orn_quat)
-
-    ret_action[0:3] = np.array([delta_pos[0], delta_pos[1], delta_orn[2]])
-
-    return ret_action
 
 
 def navigate_to_param_sampler(rng: Generator,
@@ -271,125 +155,6 @@ def create_navigate_option_model(
     return navigateToOptionModel
 
 
-def navigate_to_obj_pos(
-    env: "BehaviorEnv",
-    obj: Union["URDFObject", "RoomFloor"],
-    pos_offset: Array,
-    rng: Optional[Generator] = None
-) -> Optional[Tuple[List[List[float]], List[List[float]]]]:
-    """Parameterized controller for navigation.
-
-    Runs motion planning to find a feasible trajectory to a certain x,y
-    position offset from obj and selects an orientation such that the
-    robot is facing the object. If the navigation is infeasible, returns
-    an indication to this effect (None). Otherwise, returns the plan,
-    which is a list of list of robot base poses to move to, as well as
-    the original euler angle orientation of the robot body.
-    """
-    if rng is None:
-        rng = np.random.default_rng(23)
-
-    logging.info(f"PRIMITIVE: Attempting to navigate to {obj.name} with "
-                 f"params {pos_offset}")
-
-    # test agent positions around an obj
-    # try to place the agent near the object, and rotate it to the object
-    valid_position = None  # ((x,y,z),(roll, pitch, yaw))
-    original_orientation = p.getEulerFromQuaternion(
-        env.robots[0].get_orientation())
-    state = p.saveState()
-
-    def sample_fn(env: "BehaviorEnv",
-                  rng: Generator) -> Tuple[float, float, float]:
-        random_point = env.scene.get_random_point(rng=rng)
-        x, y = random_point[1][:2]
-        theta = (
-            rng.random() *
-            (CIRCULAR_LIMITS[1] - CIRCULAR_LIMITS[0])) + CIRCULAR_LIMITS[0]
-        return (x, y, theta)
-
-    if not isinstance(
-            obj,
-            URDFObject):  # must be a URDFObject so we can get its position!
-        logging.error("ERROR! Object to navigate to is not valid (not an "
-                      "instance of URDFObject).")
-        p.restoreState(state)
-        p.removeState(state)
-        logging.error(f"PRIMITIVE: navigate to {obj.name} with params "
-                      f"{pos_offset} fail")
-        return None
-
-    obj_pos = obj.get_position()
-    pos = [
-        pos_offset[0] + obj_pos[0],
-        pos_offset[1] + obj_pos[1],
-        env.robots[0].initial_z_offset,
-    ]
-    yaw_angle = np.arctan2(pos_offset[1], pos_offset[0]) - np.pi
-    orn = [0, 0, yaw_angle]
-    env.robots[0].set_position_orientation(pos, p.getQuaternionFromEuler(orn))
-    eye_pos = env.robots[0].parts["eye"].get_position()
-    ray_test_res = p.rayTest(eye_pos, obj_pos)
-    # Test to see if the robot is obstructed by some object, but make sure
-    # that object is not either the robot's body or the object we want to
-    # pick up!
-    blocked = len(ray_test_res) > 0 and (ray_test_res[0][0] not in (
-        env.robots[0].parts["body"].get_body_id(),
-        obj.get_body_id(),
-    ))
-    if not detect_robot_collision(env.robots[0]) and not blocked:
-        valid_position = (pos, orn)
-
-    if valid_position is None:
-        if blocked:
-            logging.warning("WARNING: Position commanded is blocked!")
-        else:
-            logging.warning("WARNING: Position commanded is in collision!")
-        p.restoreState(state)
-        p.removeState(state)
-        logging.warning(f"PRIMITIVE: navigate to {obj.name} with params "
-                        f"{pos_offset} fail")
-        return None
-
-    p.restoreState(state)
-    end_conf = [
-        valid_position[0][0],
-        valid_position[0][1],
-        valid_position[1][2],
-    ]
-    if env.use_rrt:
-        obstacles = get_body_ids(env)
-        if env.robots[0].parts["right_hand"].object_in_hand is not None:
-            obstacles.remove(env.robots[0].parts["right_hand"].object_in_hand)
-        plan = plan_base_motion_br(
-            robot=env.robots[0],
-            end_conf=end_conf,
-            base_limits=(),
-            obstacles=obstacles,
-            override_sample_fn=lambda: sample_fn(env, rng),
-            rng=rng,
-        )
-        p.restoreState(state)
-    else:
-        pos = env.robots[0].get_position()
-        plan = [[pos[0], pos[1], original_orientation[2]], end_conf]
-
-    if plan is None:
-        p.restoreState(state)
-        p.removeState(state)
-        logging.info(f"PRIMITIVE: navigate to {obj.name} with params "
-                     f"{pos_offset} failed; birrt failed to sample a plan!")
-        return None
-
-    p.restoreState(state)
-    p.removeState(state)
-
-    plan = [list(waypoint) for waypoint in plan]
-    logging.info(f"PRIMITIVE: navigate to {obj.name} success! Plan found with "
-                 f"continuous params {pos_offset}.")
-    return plan, original_orientation
-
-
 # Sampler for grasp continuous params
 def grasp_obj_param_sampler(rng: Generator) -> Array:
     """Sampler for grasp option."""
@@ -397,68 +162,6 @@ def grasp_obj_param_sampler(rng: Generator) -> Array:
     y_offset = (rng.random() * 0.4) - 0.2
     z_offset = rng.random() * 0.2
     return np.array([x_offset, y_offset, z_offset])
-
-
-def get_delta_low_level_hand_action(
-    body: "BRBody",
-    old_pos: Union[Sequence[float], Array],
-    old_orn: Union[Sequence[float], Array],
-    new_pos: Union[Sequence[float], Array],
-    new_orn: Union[Sequence[float], Array],
-) -> Array:
-    """Given a hand movement plan that is a series of waypoints for the hand in
-    position space, convert pairs of these points to a hand movement action in
-    velocity space.
-
-    Note that we cannot simply subtract subsequent positions because the
-    velocity action space used by BEHAVIOR is not defined in the world
-    frame, but rather in the frame of the previous position.
-    """
-    # First, convert the supplied orientations to quaternions
-    old_orn = p.getQuaternionFromEuler(old_orn)
-    new_orn = p.getQuaternionFromEuler(new_orn)
-
-    # Next, find the inverted position of the body (which we know shouldn't
-    # change, since our actions move either the body or the hand, but not
-    # both simultaneously)
-    inverted_body_new_pos, inverted_body_new_orn = p.invertTransform(
-        body.new_pos, body.new_orn)
-    # Use this to compute the new pose of the hand w.r.t the body frame
-    new_local_pos, new_local_orn = p.multiplyTransforms(
-        inverted_body_new_pos, inverted_body_new_orn, new_pos, new_orn)
-
-    # Next, compute the old pose of the hand w.r.t the body frame
-    inverted_body_old_pos = inverted_body_new_pos
-    inverted_body_old_orn = inverted_body_new_orn
-    old_local_pos, old_local_orn = p.multiplyTransforms(
-        inverted_body_old_pos, inverted_body_old_orn, old_pos, old_orn)
-
-    # The delta position is simply given by the difference between these
-    # positions
-    delta_pos = np.array(new_local_pos) - np.array(old_local_pos)
-
-    # Finally, compute the delta orientation
-    inverted_old_local_orn_pos, inverted_old_local_orn_orn = p.invertTransform(
-        [0, 0, 0], old_local_orn)
-    _, delta_orn = p.multiplyTransforms(
-        [0, 0, 0],
-        new_local_orn,
-        inverted_old_local_orn_pos,
-        inverted_old_local_orn_orn,
-    )
-
-    delta_trig_frac = 0
-    action = np.concatenate(
-        [
-            np.zeros((10), dtype=np.float32),
-            np.array(delta_pos, dtype=np.float32),
-            np.array(p.getEulerFromQuaternion(delta_orn), dtype=np.float32),
-            np.array([delta_trig_frac], dtype=np.float32),
-        ],
-        axis=0,
-    )
-
-    return action
 
 
 def create_grasp_policy(
@@ -698,258 +401,6 @@ def create_grasp_option_model(
     return graspObjectOptionModel
 
 
-def grasp_obj_at_pos(
-    env: "BehaviorEnv",
-    obj: Union["URDFObject", "RoomFloor"],
-    grasp_offset: Array,
-    rng: Optional[Generator] = None,
-) -> Optional[Tuple[List[List[float]], List[List[float]]]]:
-    """Parameterized controller for grasping.
-
-    Runs motion planning to find a feasible trajectory to a certain
-    x,y,z position offset from obj and selects an orientation such that
-    the palm is facing the object. If the grasp is infeasible, returns
-    an indication to this effect (None). Otherwise, returns the plan,
-    which is a list of list of hand poses, as well as the original euler
-    angle orientation of the hand.
-    """
-    if rng is None:
-        rng = np.random.default_rng(23)
-
-    logging.info(f"PRIMITIVE: Attempting to grasp {obj.name} with params "
-                 f"{grasp_offset}")
-
-    obj_in_hand = env.robots[0].parts["right_hand"].object_in_hand
-    # If we're holding something, fail and return None
-    if obj_in_hand is not None:
-        logging.info(f"PRIMITIVE: grasp {obj.name} fail, agent already has an "
-                     "object in hand!")
-        return None
-    reset_and_release_hand(env)  # first reset the hand's internal states
-
-    # If the object we're trying to grasp doesn't have all the attributes
-    # we'll need for assistive grasping, fail and return None
-    if not (isinstance(obj, URDFObject) and hasattr(obj, "states")
-            and object_states.AABB in obj.states):
-        logging.info(f"PRIMITIVE: grasp {obj.name} fail, no object")
-        return None
-
-    lo, hi = obj.states[object_states.AABB].get_value()
-    volume = get_aabb_volume(lo, hi)
-
-    # If the object is too big to be grasped, or bolted to its surface,
-    # fail and return None
-    if not (volume < 0.3 * 0.3 * 0.3 and
-            not obj.main_body_is_fixed):  # say we can only grasp small objects
-        logging.info(f"PRIMITIVE: grasp {obj.name} fail, too big or fixed")
-        return None
-
-    # If the object is too far away, fail and return None
-    if (np.linalg.norm(
-            np.array(obj.get_position()) -
-            np.array(env.robots[0].get_position())) > 2):
-        logging.info(f"PRIMITIVE: grasp {obj.name} fail, too far")
-        return None
-
-    # Grasping Phase 1: Compute the position and orientation of
-    # the hand based on the provided continuous parameters and
-    # try to create a plan to it.
-    obj_pos = obj.get_position()
-    x = obj_pos[0] + grasp_offset[0]
-    y = obj_pos[1] + grasp_offset[1]
-    z = obj_pos[2] + grasp_offset[2]
-    hand_x, hand_y, hand_z = (env.robots[0].parts["right_hand"].get_position())
-    minx = min(x, hand_x) - 0.5
-    miny = min(y, hand_y) - 0.5
-    minz = min(z, hand_z) - 0.5
-    maxx = max(x, hand_x) + 0.5
-    maxy = max(y, hand_y) + 0.5
-    maxz = max(z, hand_z) + 0.5
-
-    # compute the angle the hand must be in such that it can
-    # grasp the object from its current offset position
-    # This involves aligning the z-axis (in the world frame)
-    # of the hand with the vector that goes from the hand
-    # to the object. We can find the rotation matrix that
-    # accomplishes this rotation by following:
-    # https://math.stackexchange.com/questions/180418/
-    # calculate-rotation-matrix-to-align-vector-a-to-vector
-    # -b-in-3d
-    hand_to_obj_vector = np.array(grasp_offset[:3])
-    hand_to_obj_unit_vector = hand_to_obj_vector / \
-        np.linalg.norm(
-        hand_to_obj_vector
-    )
-    unit_z_vector = np.array([0.0, 0.0, -1.0])
-    # This is because we assume the hand is originally oriented
-    # so -z is coming out of the palm
-    c_var = np.dot(unit_z_vector, hand_to_obj_unit_vector)
-    if c_var not in [-1.0, 1.0]:
-        v_var = np.cross(unit_z_vector, hand_to_obj_unit_vector)
-        s_var = np.linalg.norm(v_var)
-        v_x = np.array([
-            [0, -v_var[2], v_var[1]],
-            [v_var[2], 0, -v_var[0]],
-            [-v_var[1], v_var[0], 0],
-        ])
-        R = (np.eye(3) + v_x + np.linalg.matrix_power(v_x, 2) * ((1 - c_var) /
-                                                                 (s_var**2)))
-        r = scipy.spatial.transform.Rotation.from_matrix(R)
-        euler_angles = r.as_euler("xyz")
-    else:
-        if c_var == 1.0:
-            euler_angles = np.zeros(3, dtype=float)
-        else:
-            euler_angles = np.array([0.0, np.pi, 0.0])
-
-    state = p.saveState()
-    end_conf = [
-        x,
-        y,
-        z,
-        euler_angles[0],
-        euler_angles[1],
-        euler_angles[2],
-    ]
-    if env.use_rrt:
-        # plan a motion to the pose [x, y, z, euler_angles[0],
-        # euler_angles[1], euler_angles[2]]
-        plan = plan_hand_motion_br(
-            robot=env.robots[0],
-            obj_in_hand=None,
-            end_conf=end_conf,
-            hand_limits=((minx, miny, minz), (maxx, maxy, maxz)),
-            obstacles=get_body_ids(env,
-                                   include_self=True,
-                                   include_right_hand=True),
-            rng=rng,
-        )
-        p.restoreState(state)
-    else:
-        pos = env.robots[0].parts["right_hand"].get_position()
-        plan = [[pos[0], pos[1], pos[2]] + list(
-            p.getEulerFromQuaternion(
-                env.robots[0].parts["right_hand"].get_orientation())),
-                end_conf]
-
-    # NOTE: This below line is *VERY* important after the
-    # pybullet state is restored. The hands keep an internal
-    # track of their state, and if we don't reset this
-    # state to mirror the actual pybullet state, the hand will
-    # think it's elsewhere and update incorrectly accordingly
-    env.robots[0].parts["right_hand"].set_position(
-        env.robots[0].parts["right_hand"].get_position())
-    env.robots[0].parts["left_hand"].set_position(
-        env.robots[0].parts["left_hand"].get_position())
-
-    # If RRT planning fails, fail and return None
-    if plan is None:
-        logging.info(f"PRIMITIVE: grasp {obj.name} fail, failed "
-                     f"to find plan to continuous params {grasp_offset}")
-        return None
-
-    # Grasping Phase 2: Move along the vector from the
-    # position the hand ends up in to the object and
-    # then try to grasp.
-    hand_pos = plan[-1][0:3]
-    hand_orn = plan[-1][3:6]
-    # Get the closest point on the object's bounding
-    # box at which we can try to put the hand
-    closest_point_on_aabb = get_closest_point_on_aabb(hand_pos, lo, hi)
-    delta_pos_to_obj = [
-        closest_point_on_aabb[0] - hand_pos[0],
-        closest_point_on_aabb[1] - hand_pos[1],
-        closest_point_on_aabb[2] - hand_pos[2],
-    ]
-    # we want to accomplish the motion in 25 timesteps
-    # NOTE: this is an arbitrary choice
-    delta_step_to_obj = [delta_pos / 25.0 for delta_pos in delta_pos_to_obj]
-
-    # move the hand along the vector to the object until it
-    # touches the object
-    for _ in range(25):
-        new_hand_pos = [
-            hand_pos[0] + delta_step_to_obj[0],
-            hand_pos[1] + delta_step_to_obj[1],
-            hand_pos[2] + delta_step_to_obj[2],
-        ]
-        plan.append(new_hand_pos + list(hand_orn))
-        hand_pos = new_hand_pos
-
-    p.restoreState(state)
-    p.removeState(state)
-    original_orientation = list(
-        p.getEulerFromQuaternion(
-            env.robots[0].parts["right_hand"].get_orientation()))
-
-    logging.info(f"PRIMITIVE: grasp {obj.name} success! Plan found with "
-                 f"continuous params {grasp_offset}.")
-    return plan, original_orientation
-
-
-def place_obj_plan(
-    env: "BehaviorEnv",
-    obj: Union["URDFObject", "RoomFloor"],
-    original_state: int,
-    place_rel_pos: Array,
-    rng: Optional[Generator] = None,
-) -> Optional[List[List[float]]]:
-    """Function to return an RRT plan for placing an object."""
-    if rng is None:
-        rng = np.random.default_rng(23)
-    obj_in_hand = env.scene.get_objects()[
-        env.robots[0].parts["right_hand"].object_in_hand]
-    x, y, z = np.add(place_rel_pos, obj.get_position())
-    hand_x, hand_y, hand_z = env.robots[0].parts["right_hand"].get_position()
-
-    minx = min(x, hand_x) - 1
-    miny = min(y, hand_y) - 1
-    minz = min(z, hand_z) - 0.5
-    maxx = max(x, hand_x) + 1
-    maxy = max(y, hand_y) + 1
-    maxz = max(z, hand_z) + 0.5
-
-    obstacles = get_body_ids(env, include_self=False)
-    obstacles.remove(env.robots[0].parts["right_hand"].object_in_hand)
-    end_conf = [
-        x,
-        y,
-        z + 0.2,
-        0,
-        np.pi * 7 / 6,
-        0,
-    ]
-    if env.use_rrt:
-        plan = plan_hand_motion_br(
-            robot=env.robots[0],
-            obj_in_hand=obj_in_hand,
-            end_conf=end_conf,
-            hand_limits=((minx, miny, minz), (maxx, maxy, maxz)),
-            obstacles=obstacles,
-            rng=rng,
-        )
-        p.restoreState(original_state)
-        p.removeState(original_state)
-    else:
-        pos = env.robots[0].parts["right_hand"].get_position()
-        plan = [[pos[0], pos[1], pos[2]] + list(
-            p.getEulerFromQuaternion(
-                env.robots[0].parts["right_hand"].get_orientation())),
-                end_conf]
-
-    # NOTE: This below line is *VERY* important after the
-    # pybullet state is restored. The hands keep an internal
-    # track of their state, and if we don't reset this
-    # state to mirror the actual pybullet state, the hand will
-    # think it's elsewhere and update incorrectly accordingly
-    env.robots[0].parts["right_hand"].set_position(
-        env.robots[0].parts["right_hand"].get_position())
-    env.robots[0].parts["left_hand"].set_position(
-        env.robots[0].parts["left_hand"].get_position())
-
-    return plan
-
-
 def place_ontop_obj_pos_sampler(
     obj: Union["URDFObject", "RoomFloor"],
     rng: Optional[Generator] = None,
@@ -964,7 +415,13 @@ def place_ontop_obj_pos_sampler(
     objA = obj[0]
     objB = obj[-1]
 
-    params = _ON_TOP_RAY_CASTING_SAMPLING_PARAMS
+    params = {
+        "max_angle_with_z_axis": 0.17,
+        "bimodal_stdev_fraction": 1e-6,
+        "bimodal_mean_fraction": 1.0,
+        "max_sampling_attempts": 50,
+        "aabb_offset": 0.01,
+    }
     aabb = get_aabb(objA.get_body_id())
     aabb_extent = get_aabb_extent(aabb)
 
@@ -1202,72 +659,3 @@ def create_place_option_model(
             env.step(np.zeros(env.action_space.shape))
 
     return placeOntopObjectOptionModel
-
-
-def place_ontop_obj_pos(
-    env: "BehaviorEnv",
-    obj: Union["URDFObject", "RoomFloor"],
-    place_rel_pos: Array,
-    rng: Optional[Generator] = None,
-) -> Optional[Tuple[List[List[float]], List[List[float]]]]:
-    """Parameterized controller for placeOnTop.
-
-    Runs motion planning to find a feasible trajectory to a certain
-    offset from obj and selects an orientation such that the palm is
-    facing the object. If the placement is infeasible, returns an
-    indication to this effect (None). Otherwise, returns the plan, which
-    is a list of list of hand poses, as well as the original euler angle
-    orientation of the hand.
-    """
-    if rng is None:
-        rng = np.random.default_rng(23)
-
-    try:
-        obj_in_hand = env.scene.get_objects()[
-            env.robots[0].parts["right_hand"].object_in_hand]
-        logging.info(f"PRIMITIVE: attempt to place {obj_in_hand.name} ontop "
-                     f"{obj.name} with params {place_rel_pos}")
-    except ValueError:
-        logging.info("Cannot place; either no object in hand or holding "
-                     "the object to be placed on top of!")
-        return None
-
-    # if the object in the agent's hand is None or not equal to the object
-    # passed in as an argument to this option, fail and return None
-    if not (obj_in_hand is not None and obj_in_hand != obj):
-        logging.info("Cannot place; either no object in hand or holding "
-                     "the object to be placed on top of!")
-        return None
-
-    # if the object is not a urdf object, fail and return None
-    if not isinstance(obj, URDFObject):
-        logging.info(f"PRIMITIVE: place {obj_in_hand.name} ontop "
-                     f"{obj.name} fail, too far")
-        return None
-
-    state = p.saveState()
-    # To check if object fits on place location
-    p.restoreState(state)
-    # NOTE: This below line is *VERY* important after the
-    # pybullet state is restored. The hands keep an internal
-    # track of their state, and if we don't reset their this
-    # state to mirror the actual pybullet state, the hand will
-    # think it's elsewhere and update incorrectly accordingly
-    env.robots[0].parts["right_hand"].set_position(
-        env.robots[0].parts["right_hand"].get_position())
-    env.robots[0].parts["left_hand"].set_position(
-        env.robots[0].parts["left_hand"].get_position())
-
-    plan = place_obj_plan(env, obj, state, place_rel_pos, rng=rng)
-    # If RRT planning fails, fail and return None
-    if plan is None:
-        logging.info(f"PRIMITIVE: placeOnTop {obj.name} fail, failed "
-                     f"to find plan to continuous params {place_rel_pos}")
-        return None
-
-    original_orientation = list(
-        p.getEulerFromQuaternion(
-            env.robots[0].parts["right_hand"].get_orientation()))
-    logging.info(f"PRIMITIVE: placeOnTop {obj.name} success! Plan found with "
-                 f"continuous params {place_rel_pos}.")
-    return plan, original_orientation
