+++ predicators/utils.py
@@ -179,35 +179,6 @@ def num_options_in_action_sequence(actions: Sequence[Action]) -> int:
     return num_options
 
 
-def get_aabb_volume(lo: Array, hi: Array) -> float:
-    """Simple utility function to compute the volume of an aabb.
-
-    lo refers to the minimum values of the bbox in the x, y and z axes,
-    while hi refers to the highest values. Both lo and hi must be three-
-    dimensional.
-    """
-    assert np.all(hi >= lo)
-    dimension = hi - lo
-    return dimension[0] * dimension[1] * dimension[2]
-
-
-def get_closest_point_on_aabb(xyz: List, lo: Array, hi: Array) -> List[float]:
-    """Get the closest point on an aabb from a particular xyz coordinate."""
-    assert np.all(hi >= lo)
-    closest_point_on_aabb = [0.0, 0.0, 0.0]
-    for i in range(3):
-        # if the coordinate is between the min and max of the aabb, then
-        # use that coordinate directly
-        if xyz[i] < hi[i] and xyz[i] > lo[i]:
-            closest_point_on_aabb[i] = xyz[i]
-        else:
-            if abs(xyz[i] - hi[i]) < abs(xyz[i] - lo[i]):
-                closest_point_on_aabb[i] = hi[i]
-            else:
-                closest_point_on_aabb[i] = lo[i]
-    return closest_point_on_aabb
-
-
 def entropy(p: float) -> float:
     """Entropy of a Bernoulli variable with parameter p."""
     assert 0.0 <= p <= 1.0
